A patch from Karl Pitrich "When editing Portal Pages, Drag & Drop for Portlets does not work" (https://issues.apache.org/jira/browse/OFBIZ-3889) - OFBIZ-3889
When editing Portlet Pages, either in MyPortal or PortalAdm, Drag & Drop of Portlets does not work.
Where are several Issues in the MyPortal, some have fixes, none is commited to trunk.

Problems:

    * in myportal.js, function getDestinationInformationPortlets(): DOM ID's are not handled correctly
    * in framework/webtools/widget/PortalAdmScreens.xml: required javascript + css files are not included


JLR: 
While working on this I have found other issues: the login.ftl files used in hhfacility and myportal were not tenant enabled. They are different  but by chance in myportal very close to common so I was finally able  to do both at along with  Karl's changes. Actually I wanted to do 2 commits but forget to uncheck the files while doing a second attempt to commit.

MyPortal
 * The login process in MyPortal was not tenant enabled, it's done in this commit.
 * I kept its specificity : newRegisterLogin and wonder why it has not been extended to common?

hhfacility
The login process in hhfacility was not tenant enabled, it's done in this commit. Doing so, I saw there i18n is not done, not an issue for me...







git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/specialpurpose/hhfacility@985163 13f79535-47bb-0310-9956-ffa450edef68